Change $peerdns to boolean.
Change $peerdns to boolean and validate the booleans.  This results in
'PEERDNS=no' appearing in the ifcfg files, even without specifying the
$peerdns paramter to the defines.  DNS1 and DNS2 will not appear in the
ifcfg files unless $peerdns => true.

Deal with the case where $dns2 is non-empty and $dns1 is empty making
sure that 'DNS1=' in the ifcfg file will always have a value.